refs #2872 Add reset menu to reset preferences
