+++ src/config/locales/en/translation.json
@@ -169,6 +169,9 @@
       "other": {
         "title": "Other options",
         "launch": "Launch app on login"
+      },
+      "reset": {
+        "button": "Reset preferences"
       }
     },
     "appearance": {
